CLI mod refactor
