+++ neuron/neuron.cabal
@@ -1,7 +1,7 @@
 cabal-version: 2.4
 name: neuron
 -- This version must be in sync with what's in Default.dhall
-version: 1.9.1.10
+version: 1.9.2.0
 license: AGPL-3.0-only
 copyright: 2020 Sridhar Ratnakumar
 maintainer: srid@srid.ca
@@ -186,11 +186,12 @@ library neuron-app-lib
     Options.Applicative.Extra
     Neuron.Backend
     Neuron.Reactor
-    Neuron.CLI
+    Neuron.CLI.App
     Neuron.CLI.New
     Neuron.CLI.Open
     Neuron.CLI.Search
     Neuron.CLI.Types
+    Neuron.CLI.Parser
     Neuron.Config
     Neuron.Version
     Neuron.Cache
