+++ neuron/src/exe/Main.hs
@@ -12,7 +12,7 @@ module Main where
 
 import GHC.IO.Handle (BufferMode (LineBuffering))
 import Main.Utf8 (withUtf8)
-import Neuron.CLI (run)
+import Neuron.CLI.App (run)
 import qualified Neuron.Reactor as Reactor
 import Relude
 import System.IO (hSetBuffering)
